Front end workspace

See merge request courses/2022-summer/csci-3130/projects/group16!14

+++ back-end/src/main/java/group16/backend/workspace/repository/WorkspaceRepo.java
@@ -4,7 +4,9 @@ import group16.backend.workspace.entity.Workspace;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface WorkspaceRepo extends JpaRepository<Workspace, Integer> {
+public interface WorkspaceRepo extends JpaRepository<Workspace, Long> {
 
 }
